+++ crates/starknet-server/src/main.rs
@@ -5,7 +5,7 @@ use ::server::ServerConfig;
 use clap::Parser;
 use cli::Args;
 use starknet_core::account::Account;
-use starknet_core::Starknet;
+use starknet_core::starknet::Starknet;
 use starknet_types::felt::Felt;
 use starknet_types::traits::{ToDecimalString, ToHexString};
 use tracing::info;
@@ -18,10 +18,13 @@ mod server;
 /// Configures tracing with default level INFO,
 /// If the environment variable `RUST_LOG` is set, it will be used instead.
 fn configure_tracing() {
-    let level_filter_layer =
-        EnvFilter::builder().with_default_directive(tracing::Level::INFO.into()).from_env_lossy();
+    let level_filter_layer = EnvFilter::builder()
+        .with_default_directive(tracing::Level::INFO.into())
+        .from_env_lossy();
 
-    tracing_subscriber::fmt().with_env_filter(level_filter_layer).init();
+    tracing_subscriber::fmt()
+        .with_env_filter(level_filter_layer)
+        .init();
 }
 
 fn log_predeployed_accounts(predeployed_accounts: &Vec<Account>, seed: u32, initial_balance: Felt) {
@@ -41,7 +44,10 @@ fn log_predeployed_accounts(predeployed_accounts: &Vec<Account>, seed: u32, init
 
     if !predeployed_accounts.is_empty() {
         println!();
-        println!("Initial balance of each account: {} WEI", initial_balance.to_decimal_string());
+        println!(
+            "Initial balance of each account: {} WEI",
+            initial_balance.to_decimal_string()
+        );
         println!("Seed to replicate this account sequence: {seed}");
     }
 }
