+++ rustfmt.toml
@@ -1,20 +1,2 @@
 edition = "2021"
-newline_style = "Unix"
-use_field_init_shorthand = true
-use_small_heuristics = "Max"
-use_try_shorthand = true
 max_width = 100
-
-# Unstable features below
-unstable_features = true
-version = "Two"
-comment_width = 100
-format_code_in_doc_comments = true
-format_macro_bodies = true
-format_macro_matchers = true
-format_strings = true
-imports_granularity = "Module"
-group_imports = "StdExternalCrate"
-normalize_comments = true
-normalize_doc_attributes = true
-wrap_comments = true
